+++ common/src/main/java/com/wynntils/features/user/QuickCastFeature.java
@@ -15,8 +15,8 @@ import com.wynntils.mc.utils.ItemUtils;
 import com.wynntils.mc.utils.McUtils;
 import com.wynntils.utils.StringUtils;
 import com.wynntils.wynn.event.WorldStateEvent;
-import com.wynntils.wynn.item.parsers.WynnItemMatchers;
 import com.wynntils.wynn.model.actionbar.SpellSegment;
+import com.wynntils.wynn.utils.WynnItemMatchers;
 import com.wynntils.wynn.utils.WynnUtils;
 import java.util.Arrays;
 import java.util.LinkedList;
