+++ common/src/main/java/com/wynntils/features/user/players/PreventTradesDuelsFeature.java
@@ -9,7 +9,7 @@ import com.wynntils.core.features.properties.FeatureCategory;
 import com.wynntils.core.features.properties.FeatureInfo;
 import com.wynntils.mc.event.PlayerAttackEvent;
 import com.wynntils.mc.event.PlayerInteractEvent;
-import com.wynntils.wynn.item.parsers.WynnItemMatchers;
+import com.wynntils.wynn.utils.WynnItemMatchers;
 import com.wynntils.wynn.utils.WynnPlayerUtils;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
