+++ common/src/main/java/com/wynntils/wynn/handleditems/annotators/game/GatheringToolAnnotator.java
@@ -8,8 +8,8 @@ import com.wynntils.handlers.item.ItemAnnotation;
 import com.wynntils.handlers.item.ItemAnnotator;
 import com.wynntils.utils.CappedValue;
 import com.wynntils.wynn.handleditems.items.game.GatheringToolItem;
-import com.wynntils.wynn.item.parsers.WynnItemMatchers;
 import com.wynntils.wynn.objects.profiles.ToolProfile;
+import com.wynntils.wynn.utils.WynnItemMatchers;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import net.minecraft.world.item.ItemStack;
