+++ common/src/main/java/com/wynntils/wynn/item/IngredientItemStack.java
@@ -5,13 +5,13 @@
 package com.wynntils.wynn.item;
 
 import com.wynntils.core.components.Managers;
-import com.wynntils.wynn.item.parsers.WynnItemMatchers;
 import com.wynntils.wynn.objects.profiles.ingredient.IngredientIdentificationContainer;
 import com.wynntils.wynn.objects.profiles.ingredient.IngredientItemModifiers;
 import com.wynntils.wynn.objects.profiles.ingredient.IngredientModifiers;
 import com.wynntils.wynn.objects.profiles.ingredient.IngredientProfile;
 import com.wynntils.wynn.objects.profiles.ingredient.ProfessionType;
 import com.wynntils.wynn.objects.profiles.item.IdentificationProfile;
+import com.wynntils.wynn.utils.WynnItemMatchers;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
