+++ common/src/main/java/com/wynntils/wynn/item/PowderItemStack.java
@@ -8,8 +8,8 @@ import com.wynntils.utils.MathUtils;
 import com.wynntils.utils.StringUtils;
 import com.wynntils.wynn.item.generator.PowderGenerator;
 import com.wynntils.wynn.item.generator.PowderProfile;
-import com.wynntils.wynn.item.parsers.WynnItemMatchers;
 import com.wynntils.wynn.objects.Powder;
+import com.wynntils.wynn.utils.WynnItemMatchers;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
