+++ common/src/main/java/com/wynntils/wynn/item/properties/DungeonKeyProperty.java
@@ -9,8 +9,8 @@ import com.wynntils.gui.render.TextRenderSetting;
 import com.wynntils.gui.render.TextRenderTask;
 import com.wynntils.mc.objects.CustomColor;
 import com.wynntils.wynn.item.WynnItemStack;
-import com.wynntils.wynn.item.parsers.WynnItemMatchers;
 import com.wynntils.wynn.item.properties.type.TextOverlayProperty;
+import com.wynntils.wynn.utils.WynnItemMatchers;
 import java.util.Arrays;
 import java.util.regex.Matcher;
 import java.util.stream.Collectors;
