+++ common/src/main/java/com/wynntils/wynn/item/properties/EmeraldPouchTierProperty.java
@@ -11,9 +11,9 @@ import com.wynntils.gui.render.TextRenderTask;
 import com.wynntils.mc.objects.CustomColor;
 import com.wynntils.utils.MathUtils;
 import com.wynntils.wynn.item.WynnItemStack;
-import com.wynntils.wynn.item.parsers.WynnItemMatchers;
 import com.wynntils.wynn.item.properties.type.HighlightProperty;
 import com.wynntils.wynn.item.properties.type.TextOverlayProperty;
+import com.wynntils.wynn.utils.WynnItemMatchers;
 import java.util.regex.Matcher;
 import net.minecraft.ChatFormatting;
