+++ common/src/main/java/com/wynntils/wynn/item/properties/TeleportScrollProperty.java
@@ -10,8 +10,8 @@ import com.wynntils.gui.render.TextRenderTask;
 import com.wynntils.mc.objects.CustomColor;
 import com.wynntils.mc.utils.ComponentUtils;
 import com.wynntils.wynn.item.WynnItemStack;
-import com.wynntils.wynn.item.parsers.WynnItemMatchers;
 import com.wynntils.wynn.item.properties.type.TextOverlayProperty;
+import com.wynntils.wynn.utils.WynnItemMatchers;
 import com.wynntils.wynn.utils.WynnUtils;
 import java.util.Arrays;
 import java.util.Locale;
