+++ common/src/main/java/com/wynntils/wynn/model/emeralds/EmeraldManager.java
@@ -12,9 +12,9 @@ import com.wynntils.mc.utils.ComponentUtils;
 import com.wynntils.mc.utils.ItemUtils;
 import com.wynntils.mc.utils.McUtils;
 import com.wynntils.wynn.event.WorldStateEvent;
-import com.wynntils.wynn.item.parsers.WynnItemMatchers;
 import com.wynntils.wynn.model.WorldStateManager;
 import com.wynntils.wynn.objects.EmeraldSymbols;
+import com.wynntils.wynn.utils.WynnItemMatchers;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
